pyproject.toml
[project]
name = "ingestion-pipeline-reports"
version = "0.1.0"
description = "Financial reports OCR ingestion pipeline (Bronze/Silver/Gold)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.0",
    "openai>=1.0",
    "supabase>=2.0",
    "transformers<5.0.0",
    "torch",
    "Pillow",
    "pymupdf",
    "python-dotenv",
    "jsonschema>=4.0",
    "anthropic>=0.25",
    "ipykernel>=7.2.0",
    "ipywidgets>=8.1.8",
    "pip>=26.0.1",
    "zai-sdk>=0.2.2",
]

[project.scripts]
ingest = "main:run"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.0",
]
